requirements.txt
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
click
Flask>=1.1.4
Flask-Cors==3.0.10
idna>=2.10
iniconfig==1.1.1
itsdangerous
Jinja2
MarkupSafe==2.0.1
packaging==21.3
peewee==3.14.9
Pillow==8.2.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.0.1
pyparsing==3.0.7
pytest==6.2.2
python-dotenv==0.19.2
requests>=2.24.0
six==1.16.0
toml==0.10.2
urllib3>=1.25.11
Werkzeug==1.0.1
psycopg2-binary==2.9.3
pytest-mock==3.7.0
deepdiff==5.7.0
twilio
json2html
dict2xml
firebase_admin
python-dateutil